Factor persistent-qq tests out
